Add Flur.ee client example, incomplete authz, Node v16
This example aims to demonstrate a client-side authz flow with signed
requests but currently is incomplete, because there needs to be a
transaction for the policy configuration, which would allow role-based
or access-limited queries to work.

A similar MWE can be done for Python or other languages, but in that
case, the correct cryptographic hash is required for the signing. The
official Flur.ee JavaScript/TypeScript library has this built in, and
therefore a Node.js client made sense.

Note that Node.js < 18 (along with an older _glibc_) requires explicit
installation of and setup for `fetch`. Later versions do not need this.
Also, in a true client, the private key needs to be (securely) saved and
loaded, in order to produce the same `did` every time.

Details about this security mechanism is not thoroughly documented for
Flur.ee v3 yet, but discussions towards this can be found on their
official Discord community channel, for e.g. the following thread where
a complete dummy example for authorization is given (without signing):

https://discord.com/channels/896089675511508992/1245329089343262720/1245521794589266051

Please also ensure, if not using the hosted Flur.ee service, that the
correct Docker image `server` is used, and not `ledger`, which is a v2
thing. As per current documentation, the following command should be
used to run Flur.ee locally via Docker:

```
docker run -p 58090:8090 -v `pwd`/data:/opt/fluree-server/data fluree/server
```

You may need to create the `data` directory beforehand if you get any
errors starting up (or create a suitable directory elsewhere and change
the first part of the `-v` mount argument).

Always refer to the latest docs
https://next.developers.flur.ee/docs/learn/tutorial/introduction/ (the
URL should _not_ start with _docsarchive_). The end of the official
cookbook as some limited documentation on security:

https://next.developers.flur.ee/docs/reference/cookbook/#policy